Seems to work!
Checked all examples.



git-svn-id: svn://136.177.114.72/svn_GW/phreeqc/branches/cpp@729 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ read.cpp
@@ -6214,7 +6214,7 @@ int check_units (char *tot_units,  int alkalinity, int check_compatibility,
  *   Check if unit in list
  */
 	found = FALSE;
-	for (i=0; i < NUNITS; i++) {
+	for (i=0; i < (int) NUNITS; i++) {
 		if (strcmp(tot_units, units[i]) == 0) {
 			found=TRUE;
 			break;
